refactor(app): :art: Consolidation of CLI and GUI shutdown logic (reduces code dupe)
